requirements.txt
streamlit>=1.20
sentence-transformers>=2.2.2
scikit-learn
pdfplumber
python-docx
spacy>=3.0
reportlab
openai
rapidfuzz    # optional but strongly recommended
fastapi
uvicorn
python-multipart
openai
python-jose
passlib[bcrypt]
faiss-cpu
sentence-transformers
streamlit==1.36.0
pdfplumber==0.10.3
python-docx==1.1.2
sentence-transformers==2.6.1
scikit-learn==1.5.0
numpy==1.26.4
spacy==3.7.2
reportlab==4.2.0
pandas==2.2.2
